prevent compiler warning

+++ src/AsciiFile.cc
@@ -163,7 +163,8 @@ bool AsciiFile::updateFile()
 		access( OldBakName_Ci.c_str(), R_OK ) != 0 )
 		{
 		int r = link( BakName_Ci.c_str(), OldBakName_Ci.c_str() );
-		unlink( BakName_Ci.c_str() );
+		if( r==0 )
+		    unlink( BakName_Ci.c_str() );
 		}
 	    SystemCmd Cmd_Ci;
 	    string Tmp_Ci = "cp -a ";
